jittor/requirements.txt
jittor==1.3.9.14
numpy==1.26.4
hydra-core==1.3.2
omegaconf==2.3.0
opencv-python==4.11.0.86
matplotlib==3.10.3